Refactor (FolderSelect, CardList) : 컴포넌트 분리
