+++ components/Header/Header.tsx
@@ -5,19 +5,16 @@ import { URLS } from "@/utils/getData.type";
 
 export default function Header() {
   const folderName = useData(URLS.SHARED_FOLDERNAME);
+  if (!folderName) {
+    return null;
+  }
 
   return (
     <Container>
       <User>
-        {folderName ? (
-          <>
-            <UserImg src={folderName.owner.profileImageSource ?? defaultAvatar} alt="유저 프로필 이미지" />
-            <UserText>{folderName.owner.name}</UserText>
-            <UserTitle>{folderName.folderName}</UserTitle>
-          </>
-        ) : (
-          <p>폴더 정보를 읽어오는 데 실패했습니다.</p>
-        )}
+        <UserImg src={folderName.owner.profileImageSource ?? defaultAvatar.src} alt="유저 프로필 이미지" />
+        <UserText>{folderName.owner.name}</UserText>
+        <UserTitle>{folderName.folderName}</UserTitle>
       </User>
     </Container>
   );
