+++ components/Main/FolderSelect/SearchBar.styled.tsx
