+++ components/Nav/Avatar/Profile.tsx
@@ -1,4 +1,5 @@
 import { Container, ProfileImg, ProfileText } from "@/components/Nav/Avatar/Profile.styled";
+import defaultProfileImg from "@/public/Avatar.png";
 
 interface Props {
   profileImg: string;
@@ -8,7 +9,7 @@ interface Props {
 export default function Profile({ profileImg, email }: Props) {
   return (
     <Container>
-      <ProfileImg src={profileImg} alt="프로필 사진" />
+      <ProfileImg src={profileImg ?? defaultProfileImg.src} alt="프로필 사진" />
       <ProfileText>{email}</ProfileText>
     </Container>
   );
